new benchmark tests added
